Extract chain details form into separate component
